Push tests out to new file for fixes
